// frontend/src/features/common/types/skill.ts
import { StrapiMediaResponse } from "./media";

export interface Skill {
  id: number;
  name: string;
  category: string;
  proficiency: number;
  description?: string;
  order: number;
  visible?: boolean;
  isPublic?: boolean;
  icon?: StrapiMediaResponse;
}

export interface SkillsResponse {
  data: Skill[];
} 